pyproject.toml
[tool.poetry]
name = "be-course-app"
version = "0.1.0"
description = ""
authors = ["arystambek-dimash <arystambekdimash005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pydantic-settings = "^2.6.0"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
psycopg-binary = "^3.2.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
